backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic_settings
python-dotenv
alembic
uuid
beautifulsoup4
requests
python-dateutil
scikit-learn